+++ crates/swc_plugin_runner/src/imported_fn/metadata_context.rs
@@ -5,18 +5,16 @@ use swc_common::plugin::{
     metadata::{TransformPluginMetadataContext, TransformPluginMetadataContextKind},
     serialized::PluginSerializedBytes,
 };
-use wasmer::{LazyInit, Memory, NativeFunc};
+use wasmer::{AsStoreMut, FunctionEnvMut, Memory, TypedFunction};
 
 use crate::memory_interop::{allocate_return_values_into_guest, copy_bytes_into_host};
 
-#[derive(wasmer::WasmerEnv, Clone)]
+#[derive(Clone)]
 pub struct MetadataContextHostEnvironment {
-    #[wasmer(export)]
-    pub memory: wasmer::LazyInit<Memory>,
+    pub memory: Option<Memory>,
     /// Attached imported fn `__alloc` to the hostenvironment to allow any other
     /// imported fn can allocate guest's memory space from host runtime.
-    #[wasmer(export(name = "__alloc"))]
-    pub alloc_guest_memory: LazyInit<NativeFunc<u32, i32>>,
+    pub alloc_guest_memory: Option<TypedFunction<i32, i32>>,
     pub metadata_context: Arc<TransformPluginMetadataContext>,
     pub transform_plugin_config: Option<serde_json::Value>,
     /// A buffer to string key to the context plugin need to pass to the host.
@@ -30,8 +28,8 @@ impl MetadataContextHostEnvironment {
         mutable_context_key_buffer: &Arc<Mutex<Vec<u8>>>,
     ) -> Self {
         MetadataContextHostEnvironment {
-            memory: LazyInit::default(),
-            alloc_guest_memory: LazyInit::default(),
+            memory: None,
+            alloc_guest_memory: None,
             metadata_context: metadata_context.clone(),
             transform_plugin_config: plugin_config.clone(),
             mutable_context_key_buffer: mutable_context_key_buffer.clone(),
@@ -43,24 +41,24 @@ impl MetadataContextHostEnvironment {
 /// in the host can read it.
 #[tracing::instrument(level = "info", skip_all)]
 pub fn copy_context_key_to_host_env(
-    env: &MetadataContextHostEnvironment,
+    mut env: FunctionEnvMut<MetadataContextHostEnvironment>,
     bytes_ptr: i32,
     bytes_ptr_len: i32,
 ) {
-    if let Some(memory) = env.memory_ref() {
-        (*env.mutable_context_key_buffer.lock()) =
-            copy_bytes_into_host(memory, bytes_ptr, bytes_ptr_len);
+    if let Some(memory) = env.data().memory.as_ref() {
+        (*env.data_mut().mutable_context_key_buffer.lock()) =
+            copy_bytes_into_host(&memory.view(&env), bytes_ptr, bytes_ptr_len);
     }
 }
 
 #[tracing::instrument(level = "info", skip_all)]
 pub fn get_transform_plugin_config(
-    env: &MetadataContextHostEnvironment,
+    mut env: FunctionEnvMut<MetadataContextHostEnvironment>,
     allocated_ret_ptr: i32,
 ) -> i32 {
-    if let Some(memory) = env.memory_ref() {
-        if let Some(alloc_guest_memory) = env.alloc_guest_memory_ref() {
-            let config_value = &env.transform_plugin_config;
+    if let Some(memory) = env.data().memory.clone().as_ref() {
+        if let Some(alloc_guest_memory) = env.data().alloc_guest_memory.clone().as_ref() {
+            let config_value = &env.data().transform_plugin_config;
             if let Some(config_value) = config_value {
                 // Lazy as possible as we can - only deserialize json value if transform plugin
                 // actually needs it.
@@ -71,6 +69,7 @@ pub fn get_transform_plugin_config(
 
                     allocate_return_values_into_guest(
                         memory,
+                        &mut env.as_store_mut(),
                         alloc_guest_memory,
                         allocated_ret_ptr,
                         &serialized,
@@ -86,13 +85,17 @@ pub fn get_transform_plugin_config(
 
 #[tracing::instrument(level = "info", skip_all)]
 pub fn get_transform_context(
-    env: &MetadataContextHostEnvironment,
+    mut env: FunctionEnvMut<MetadataContextHostEnvironment>,
     key: u32,
     allocated_ret_ptr: i32,
 ) -> i32 {
-    if let Some(memory) = env.memory_ref() {
-        if let Some(alloc_guest_memory) = env.alloc_guest_memory_ref() {
+    let memory = env.data().memory.clone();
+    let __alloc = env.data().alloc_guest_memory.clone();
+
+    if let Some(memory) = memory.as_ref() {
+        if let Some(alloc_guest_memory) = __alloc.as_ref() {
             let value = env
+                .data()
                 .metadata_context
                 .get(&TransformPluginMetadataContextKind::from(key));
 
@@ -102,6 +105,7 @@ pub fn get_transform_context(
 
                 allocate_return_values_into_guest(
                     memory,
+                    &mut env.as_store_mut(),
                     alloc_guest_memory,
                     allocated_ret_ptr,
                     &serialized,
@@ -116,17 +120,18 @@ pub fn get_transform_context(
 
 #[tracing::instrument(level = "info", skip_all)]
 pub fn get_experimental_transform_context(
-    env: &MetadataContextHostEnvironment,
+    mut env: FunctionEnvMut<MetadataContextHostEnvironment>,
     allocated_ret_ptr: i32,
 ) -> i32 {
-    if let Some(memory) = env.memory_ref() {
-        if let Some(alloc_guest_memory) = env.alloc_guest_memory_ref() {
-            let context_key_buffer = &*env.mutable_context_key_buffer.lock();
+    if let Some(memory) = env.data().memory.clone().as_ref() {
+        if let Some(alloc_guest_memory) = env.data().alloc_guest_memory.clone().as_ref() {
+            let context_key_buffer = env.data().mutable_context_key_buffer.lock().clone();
             let key: String = PluginSerializedBytes::from_slice(&context_key_buffer[..])
                 .deserialize()
                 .expect("Should able to deserialize");
 
             let value = env
+                .data()
                 .metadata_context
                 .experimental
                 .get(&key)
@@ -138,6 +143,7 @@ pub fn get_experimental_transform_context(
 
                 allocate_return_values_into_guest(
                     memory,
+                    &mut env.as_store_mut(),
                     alloc_guest_memory,
                     allocated_ret_ptr,
                     &serialized,
@@ -152,17 +158,18 @@ pub fn get_experimental_transform_context(
 
 #[tracing::instrument(level = "info", skip_all)]
 pub fn get_raw_experiemtal_transform_context(
-    env: &MetadataContextHostEnvironment,
+    mut env: FunctionEnvMut<MetadataContextHostEnvironment>,
     allocated_ret_ptr: i32,
 ) -> i32 {
-    if let Some(memory) = env.memory_ref() {
-        let experimental_context = &env.metadata_context.experimental;
+    if let Some(memory) = env.data().memory.clone().as_ref() {
+        let experimental_context = env.data().metadata_context.experimental.clone();
         let serialized_experimental_context_bytes =
-            PluginSerializedBytes::try_serialize(experimental_context)
+            PluginSerializedBytes::try_serialize(&experimental_context)
                 .expect("Should be serializable");
-        if let Some(alloc_guest_memory) = env.alloc_guest_memory_ref() {
+        if let Some(alloc_guest_memory) = env.data().alloc_guest_memory.clone().as_ref() {
             allocate_return_values_into_guest(
                 memory,
+                &mut env.as_store_mut(),
                 alloc_guest_memory,
                 allocated_ret_ptr,
                 &serialized_experimental_context_bytes,