+++ crates/swc_plugin_runner/src/imported_fn/source_map.rs
@@ -6,20 +6,18 @@ use swc_common::{
     source_map::{PartialFileLines, PartialLoc},
     BytePos, SourceMap, Span, SyntaxContext,
 };
-use wasmer::{LazyInit, Memory, NativeFunc};
+use wasmer::{AsStoreMut, FunctionEnvMut, Memory, TypedFunction};
 
 use crate::memory_interop::{allocate_return_values_into_guest, write_into_memory_view};
 
 /// External environment state for imported (declared in host, injected into
 /// guest) fn for source map proxy.
-#[derive(wasmer::WasmerEnv, Clone)]
+#[derive(Clone)]
 pub struct SourceMapHostEnvironment {
-    #[wasmer(export)]
-    pub memory: wasmer::LazyInit<Memory>,
+    pub memory: Option<Memory>,
     /// Attached imported fn `__alloc` to the hostenvironment to allow any other
     /// imported fn can allocate guest's memory space from host runtime.
-    #[wasmer(export(name = "__alloc"))]
-    pub alloc_guest_memory: LazyInit<NativeFunc<u32, i32>>,
+    pub allocate_guest_memory: Option<TypedFunction<i32, i32>>,
     pub source_map: Arc<Mutex<Arc<SourceMap>>>,
     /// A buffer to non-determined size of return value from the host.
     pub mutable_source_map_buffer: Arc<Mutex<Vec<u8>>>,
@@ -31,8 +29,8 @@ impl SourceMapHostEnvironment {
         mutable_source_map_buffer: &Arc<Mutex<Vec<u8>>>,
     ) -> SourceMapHostEnvironment {
         SourceMapHostEnvironment {
-            memory: LazyInit::default(),
-            alloc_guest_memory: LazyInit::default(),
+            memory: None,
+            allocate_guest_memory: None,
             source_map: source_map.clone(),
             mutable_source_map_buffer: mutable_source_map_buffer.clone(),
         }
@@ -47,13 +45,13 @@ impl SourceMapHostEnvironment {
 /// `SourceFile` is needed.
 #[tracing::instrument(level = "info", skip_all)]
 pub fn lookup_char_pos_proxy(
-    env: &SourceMapHostEnvironment,
+    mut env: FunctionEnvMut<SourceMapHostEnvironment>,
     byte_pos: u32,
     should_include_source_file: i32,
     allocated_ret_ptr: i32,
 ) -> i32 {
-    if let Some(memory) = env.memory_ref() {
-        let original_loc = (env.source_map.lock()).lookup_char_pos(BytePos(byte_pos));
+    if let Some(memory) = env.data().memory.clone().as_ref() {
+        let original_loc = (env.data().source_map.lock()).lookup_char_pos(BytePos(byte_pos));
         let ret = PartialLoc {
             source_file: if should_include_source_file == 0 {
                 None
@@ -68,9 +66,10 @@ pub fn lookup_char_pos_proxy(
         let serialized_loc_bytes =
             PluginSerializedBytes::try_serialize(&ret).expect("Should be serializable");
 
-        if let Some(alloc_guest_memory) = env.alloc_guest_memory_ref() {
+        if let Some(alloc_guest_memory) = env.data().allocate_guest_memory.clone().as_ref() {
             allocate_return_values_into_guest(
                 memory,
+                &mut env.as_store_mut(),
                 alloc_guest_memory,
                 allocated_ret_ptr,
                 &serialized_loc_bytes,
@@ -85,14 +84,14 @@ pub fn lookup_char_pos_proxy(
 }
 
 #[tracing::instrument(level = "info", skip_all)]
-pub fn doctest_offset_line_proxy(env: &SourceMapHostEnvironment, orig: u32) -> u32 {
-    (env.source_map.lock()).doctest_offset_line(orig as usize) as u32
+pub fn doctest_offset_line_proxy(env: FunctionEnvMut<SourceMapHostEnvironment>, orig: u32) -> u32 {
+    (env.data().source_map.lock()).doctest_offset_line(orig as usize) as u32
 }
 
 #[allow(clippy::too_many_arguments)]
 #[tracing::instrument(level = "info", skip_all)]
 pub fn merge_spans_proxy(
-    env: &SourceMapHostEnvironment,
+    mut env: FunctionEnvMut<SourceMapHostEnvironment>,
     lhs_lo: u32,
     lhs_hi: u32,
     lhs_ctxt: u32,
@@ -101,7 +100,7 @@ pub fn merge_spans_proxy(
     rhs_ctxt: u32,
     allocated_ptr: i32,
 ) -> i32 {
-    if let Some(memory) = env.memory_ref() {
+    if let Some(memory) = env.data().memory.clone().as_ref() {
         let sp_lhs = Span {
             lo: BytePos(lhs_lo),
             hi: BytePos(lhs_hi),
@@ -114,11 +113,16 @@ pub fn merge_spans_proxy(
             ctxt: SyntaxContext::from_u32(rhs_ctxt),
         };
 
-        let ret = (env.source_map.lock()).merge_spans(sp_lhs, sp_rhs);
+        let ret = (env.data().source_map.lock()).merge_spans(sp_lhs, sp_rhs);
         if let Some(span) = ret {
             let serialized_bytes =
                 PluginSerializedBytes::try_serialize(&span).expect("Should be serializable");
-            write_into_memory_view(memory, &serialized_bytes, |_| allocated_ptr);
+            write_into_memory_view(
+                memory,
+                &mut env.as_store_mut(),
+                &serialized_bytes,
+                |_, _| allocated_ptr,
+            );
             1
         } else {
             0
@@ -130,21 +134,21 @@ pub fn merge_spans_proxy(
 
 #[tracing::instrument(level = "info", skip_all)]
 pub fn span_to_lines_proxy(
-    env: &SourceMapHostEnvironment,
+    mut env: FunctionEnvMut<SourceMapHostEnvironment>,
     span_lo: u32,
     span_hi: u32,
     span_ctxt: u32,
     should_request_source_file: i32,
     allocated_ret_ptr: i32,
 ) -> i32 {
-    if let Some(memory) = env.memory_ref() {
+    if let Some(memory) = env.data().memory.clone().as_ref() {
         let span = Span {
             lo: BytePos(span_lo),
             hi: BytePos(span_hi),
             ctxt: SyntaxContext::from_u32(span_ctxt),
         };
 
-        let ret = (env.source_map.lock())
+        let ret = (env.data().source_map.lock())
             .span_to_lines(span)
             .map(|lines| PartialFileLines {
                 file: if should_request_source_file == 0 {
@@ -158,9 +162,10 @@ pub fn span_to_lines_proxy(
         let serialized_loc_bytes =
             PluginSerializedBytes::try_serialize(&ret).expect("Should be serializable");
 
-        if let Some(alloc_guest_memory) = env.alloc_guest_memory_ref() {
+        if let Some(alloc_guest_memory) = env.data().allocate_guest_memory.clone().as_ref() {
             allocate_return_values_into_guest(
                 memory,
+                &mut env.as_store_mut(),
                 alloc_guest_memory,
                 allocated_ret_ptr,
                 &serialized_loc_bytes,
@@ -176,20 +181,21 @@ pub fn span_to_lines_proxy(
 
 #[tracing::instrument(level = "info", skip_all)]
 pub fn lookup_byte_offset_proxy(
-    env: &SourceMapHostEnvironment,
+    mut env: FunctionEnvMut<SourceMapHostEnvironment>,
     byte_pos: u32,
     allocated_ret_ptr: i32,
 ) -> i32 {
-    if let Some(memory) = env.memory_ref() {
+    if let Some(memory) = env.data().memory.clone().as_ref() {
         let byte_pos = BytePos(byte_pos);
-        let ret = (env.source_map.lock()).lookup_byte_offset(byte_pos);
+        let ret = (env.data().source_map.lock()).lookup_byte_offset(byte_pos);
 
         let serialized_loc_bytes =
             PluginSerializedBytes::try_serialize(&ret).expect("Should be serializable");
 
-        if let Some(alloc_guest_memory) = env.alloc_guest_memory_ref() {
+        if let Some(alloc_guest_memory) = env.data().allocate_guest_memory.clone().as_ref() {
             allocate_return_values_into_guest(
                 memory,
+                &mut env.as_store_mut(),
                 alloc_guest_memory,
                 allocated_ret_ptr,
                 &serialized_loc_bytes,
@@ -205,25 +211,26 @@ pub fn lookup_byte_offset_proxy(
 
 #[tracing::instrument(level = "info", skip_all)]
 pub fn span_to_string_proxy(
-    env: &SourceMapHostEnvironment,
+    mut env: FunctionEnvMut<SourceMapHostEnvironment>,
     span_lo: u32,
     span_hi: u32,
     span_ctxt: u32,
     allocated_ret_ptr: i32,
 ) -> i32 {
-    if let Some(memory) = env.memory_ref() {
+    if let Some(memory) = env.data().memory.clone().as_ref() {
         let span = Span {
             lo: BytePos(span_lo),
             hi: BytePos(span_hi),
             ctxt: SyntaxContext::from_u32(span_ctxt),
         };
-        let ret = (env.source_map.lock()).span_to_string(span);
+        let ret = (env.data().source_map.lock()).span_to_string(span);
         let serialized_loc_bytes =
             PluginSerializedBytes::try_serialize(&ret).expect("Should be serializable");
 
-        if let Some(alloc_guest_memory) = env.alloc_guest_memory_ref() {
+        if let Some(alloc_guest_memory) = env.data().allocate_guest_memory.clone().as_ref() {
             allocate_return_values_into_guest(
                 memory,
+                &mut env.as_store_mut(),
                 alloc_guest_memory,
                 allocated_ret_ptr,
                 &serialized_loc_bytes,
@@ -239,25 +246,26 @@ pub fn span_to_string_proxy(
 
 #[tracing::instrument(level = "info", skip_all)]
 pub fn span_to_filename_proxy(
-    env: &SourceMapHostEnvironment,
+    mut env: FunctionEnvMut<SourceMapHostEnvironment>,
     span_lo: u32,
     span_hi: u32,
     span_ctxt: u32,
     allocated_ret_ptr: i32,
 ) -> i32 {
-    if let Some(memory) = env.memory_ref() {
+    if let Some(memory) = env.data().memory.clone().as_ref() {
         let span = Span {
             lo: BytePos(span_lo),
             hi: BytePos(span_hi),
             ctxt: SyntaxContext::from_u32(span_ctxt),
         };
-        let ret = (env.source_map.lock()).span_to_filename(span);
+        let ret = (env.data().source_map.lock()).span_to_filename(span);
         let serialized_loc_bytes =
             PluginSerializedBytes::try_serialize(&ret).expect("Should be serializable");
 
-        if let Some(alloc_guest_memory) = env.alloc_guest_memory_ref() {
+        if let Some(alloc_guest_memory) = env.data().allocate_guest_memory.clone().as_ref() {
             allocate_return_values_into_guest(
                 memory,
+                &mut env.as_store_mut(),
                 alloc_guest_memory,
                 allocated_ret_ptr,
                 &serialized_loc_bytes,